Rename perl5 program and re-run timings
Want to make all the scripts have the same name, and the elixir ends up
being called `look_and_say`.  That makes it easier to run the scripts by
just changing the directory name (e.g. from `perl5/` to `elixir1/`.

Re-ran the timings with my laptop plugged in so that the conditions are
the same for all runs in terms of clock speed etc.